requirements.txt
streamlit
langchain
faiss-cpu
pypdf
sentence-transformers
googletrans==4.0.0-rc1
openai
python-dotenv